move RevList functionality to git package
